% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frames.R
\name{as.EprimeFrame}
\alias{as.EprimeFrame}
\title{Convert a list into an EprimeFrame object}
\usage{
as.EprimeFrame(xs)
}
\arguments{
\item{xs}{a list}
}
\value{
the original list as an \code{EprimeFrame} object (along with dummy
  Eprime metadata fields)
}
\description{
Convert a list into an EprimeFrame object
}

