% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.FrameList}
\alias{as.FrameList}
\title{Convert a list of EprimeFrames into a FrameList object}
\usage{
as.FrameList(xs)
}
\arguments{
\item{xs}{a list of EprimeFrames}
}
\value{
the original list as a \code{FrameList} object
}
\description{
Convert a list of EprimeFrames into a FrameList object
}

