% Generated by roxygen2 (4.0.2): do not edit by hand
\name{count_tabs}
\alias{count_tabs}
\title{Infer level of nesting (for a log-frame) by counting tabs}
\usage{
count_tabs(colon_sep_xs)
}
\description{
The number of tabs before the "key: value" information in a log-frame tells
where the frame is nested in the experiment's structure.
}
\keyword{internal}

