% Generated by roxygen2 (4.0.1): do not edit by hand
\name{EprimeFrame}
\alias{EprimeFrame}
\title{Create an EprimeFrame object}
\usage{
EprimeFrame(keys_values)
}
\arguments{
\item{keys_values}{}
}
\value{
a list with the class \code{EprimeFrame} and with special
  \code{Eprime.} metadata, \code{Running} and \code{Procedure} values, all
  set to NA by default.
}
\description{
This constructor function converts a character vector object into an
\code{EprimeFrame} object, which is just a list with some special metadata
values.
}

