% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extract_chunks}
\alias{extract_chunks}
\title{Extract chunks of text between Eprime LogFrame boundaries.}
\usage{
extract_chunks(eprime_log, metadata = TRUE)
}
\arguments{
\item{eprime_log}{a character vector containing the lines of text from Eprime
txt file}

\item{metadata}{whether to include the additional lines of metadata. TRUE by
default.}
}
\value{
a list of character vectors, each vector containing the lines of text
  that fall between Eprime logframe boundaries.
}
\description{
Almost all of information in an Eprime file comes in chunks of text bracketed
by the lines \code{*** LogFrame Start ***} and \code{*** LogFrame End ***}.
The exception is the header information which is bracketed by \code{***
Header Start ***} and \code{*** Header End ***}.
}
\details{
\code{extract_chunks} extracts the bracketed text, storing each chunk of text
in a list. By default, the lists also include additional lines of text as
metadata: \code{Eprime.FrameNumber} and \code{Eprime.Basename} (the name of
the source file). The header frame also gets dummy lines: \code{Procedure:
Header} and \code{Running: Header}.
}

