% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extract_frames}
\alias{extract_frames}
\title{Extract chunks of text and convert into Eprime Frames}
\usage{
extract_frames(eprime_log)
}
\arguments{
\item{eprime_log}{a character vector containing the lines of text from Eprime
txt file}
}
\value{
a FrameList object (a list of EprimeFrame objects)
}
\description{
This is shortcut for \code{make_eprime_frames(extract_chunks(eprime_log))}
}

