% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_in}
\alias{filter_in}
\alias{filter_out}
\title{Filter levels in or out of a FrameList based on attribute values}
\usage{
filter_in(frame_list, key, values)

filter_out(frame_list, key, values)
}
\arguments{
\item{frame_list}{a list of \code{EprimeFrame} objects}

\item{key}{the name of the attribute to filter in or out}

\item{values}{the whitelisted or blacklisted values of the attribute}
}
\value{
for \code{filter_in}, only log-frames where \code{key} is one of the
  \code{values} are kept. for \code{filter_out}, log-frames where \code{key}
  is one of the \code{values} are omitted.
}
\description{
Filter levels in or out of a FrameList based on attribute values
}
