% Generated by roxygen2 (4.0.1): do not edit by hand
\name{keep_levels}
\alias{drop_levels}
\alias{keep_levels}
\title{Filter levels in or out of a FrameList based on Eprime.Level values}
\usage{
keep_levels(frame_list, level_numbers)

drop_levels(frame_list, level_numbers)
}
\arguments{
\item{level_numbers}{the whitelisted or blacklisted values for Eprime.Level}

\item{frame_list}{a list of \code{EprimeFrame} objects}
}
\value{
for \code{keep_levels}, only log-frames where the level matches one
  of the \code{level_numbers} are kept. for \code{keep_levels}, log-frames
  where the level matches one of the \code{level_numbers} are omitted.
}
\description{
These functions are shortcuts for calls to \code{filter_in} or
\code{filter_out}.
}
\details{
Note that the meaning of Eprime.Level value in a log-frame ultimately is
equal to one plus the number of tabs before each line in the log-frame.
}

