% Generated by roxygen2 (4.0.1): do not edit by hand
\name{list_functions}
\alias{list_functions}
\alias{omit}
\alias{omit_apply}
\alias{pick}
\alias{pick_apply}
\alias{pluck}
\alias{pluck_apply}
\title{Higher-order functions for dealing with lists}
\usage{
pluck(key)

pluck_apply(xss, key)

pick(keys)

pick_apply(xss, keys)

omit(keys)

omit_apply(xss, keys)
}
\arguments{
\item{key}{the name of the value to pluck from a list}

\item{keys}{a character vector of names used to simplify the list}

\item{xs}{a list}

\item{xss}{a list of lists

\itemize{
  \item \code{pluck}: Pluck a named value from a list
  \item \code{pick}: Simplify a list by picking out whitelisted names
  \item \code{omit}: Simplify a list by omitting blacklisted names.
}}
}
\description{
These functions were inspired by underscore.js.
}

