% Generated by roxygen2 (4.0.1): do not edit by hand
\name{listify}
\alias{listify}
\title{Convert a vector of colon-separated text lines into a list of named elements}
\usage{
listify(colon_sep_values)
}
\arguments{
\item{colon_sep_values}{a character vector with lines of the form \code{"key:
value"}}
}
\value{
a named list of the values in the colon-separated lines. \code{"key:
  value"} yields \code{list(key = "value")}
}
\description{
Convert a vector of colon-separated text lines into a list of named elements
}
\details{
Some minoring cleaning of the input is performed:
  \itemize{
    \item White-space on both sides of the input strings is omitted.
    \item Eprime bracketing lines like \code{"*** LogFrame Start ***"}
          are filtered out.
    \item Lines without a colon character \code{":"} are filtered out.}
}
\examples{
lines <- c("\\t*** LogFrame Start ***",
           "\\tProcedure: PracticeProc",
           "\\tStimulus1: toothbrush",
           "\\t*** LogFrame End ***")

str(listify(lines))
# List of 2
# $ Procedure: chr "PracticeProc"
# $ Stimulus1: chr "toothbrush"

lines2 <- c("Item: 1",
            "  WhiteSpaceItem: Stuff\\t",
            "AnotherItem: Two Colons: Yes",
            "",
            "Last Item: Whatever")

str(listify(lines2))
# List of 4
# $ Item          : chr "1"
# $ WhiteSpaceItem: chr "Stuff"
# $ AnotherItem   : chr "Two Colons: Yes"
# $ Last Item     : chr "Whatever"
}

