% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extract.R
\name{make_ranges}
\alias{make_ranges}
\title{Check Eprime log-frame line ranges}
\usage{
make_ranges(starts, ends, eprime_log)
}
\arguments{
\item{starts}{the line numbers of the log-frame start lines}

\item{ends}{the line numbers of the log-frame end lines}

\item{eprime_log}{a character vector of lines from an Eprime log file}
}
\value{
a list of sequences. Each list contains the line numbers contained by
  an Eprime log-frame. If there is a log frame without an end-line, that
  partial frame is excluded and its contents are previed in a warning
  message.
}
\description{
Check Eprime log-frame line ranges
}
\keyword{internal}

