% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extract.R
\name{parse_chunks}
\alias{parse_chunks}
\title{Extract the lines of text from each log-frame}
\usage{
parse_chunks(eprime_log)
}
\arguments{
\item{eprime_log}{a character vector of lines from an Eprime log file}
}
\value{
a list of character vectors, where each vector contains the lines of
  a log-frame
}
\description{
Extract the lines of text from each log-frame
}
\keyword{internal}

