% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read_eprime}
\alias{read_eprime}
\title{Read in a text file generated by Eprime}
\usage{
read_eprime(filename)
}
\arguments{
\item{filename}{Either the full or relative path to the .txt file that is to
be parsed.}
}
\value{
Each line of the file is stored and returned in a character vector.
}
\description{
Read in a text file generated by Eprime
}
\details{
The encoding on an Eprime txt file should be UCS-2 Little Endian,
  but sometimes this is not the case. We delegate the fussy encoding details
  to the \code{stringi::str_read_lines} function.
}

