% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rprime-package.R
\docType{package}
\name{rprime}
\alias{rprime-package}
\alias{rprime}
\title{Functions for dealing with Eprime txt files}
\description{
'Eprime' is a set of programs for administering psychological experiments by computer. This package provides functions for loading, parsing, filtering and exporting data in the text files produced by 'Eprime' experiments.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tjmahr/rprime}
  \item Report bugs at \url{https://github.com/tjmahr/rprime/issues}
}

}
\author{
\strong{Maintainer}: Tristan Mahr \email{tristan.mahr@wisc.edu} (\href{https://orcid.org/0000-0002-8890-5116}{ORCID})

}
\keyword{internal}
