% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tidy}
\alias{tidy}
\title{Clean up `Running`-related attributes}
\usage{
tidy(x)
}
\description{
A fresh EprimeFrame object has the structure:
}
\details{
\code{
  Eprime.LevelName: NA
  Eprime.Level: [Level]
  Running: [Key]
  [Key]: [Value]
  [Key].Cycle: [Cycle]
  [Key].Sample: [Sample]
}

These \code{[Key]} values make it harder to merge together data-frames later
on, since each unique \code{[Key]} gets its own column name. Therefore, we
normalize these field names early on. The end result has the structure:

\code{
  Eprime.LevelName: [Key]_[Level]
  Eprime.Level: [Level]
  Running: [Key]
  Cycle: [Cycle]
  Sample: [Sample]
}
}
\keyword{internal}

