% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_data_frame.R
\name{to_data_frame}
\alias{to_data_frame}
\title{Convert Eprime Frames into data-frames}
\usage{
to_data_frame(x)
}
\arguments{
\item{x}{an EprimeFrame object, or a FrameList object (a list of
EprimeFrames)}
}
\value{
all of the EprimeFrames combined into a single data frame.
}
\description{
Convert Eprime Frames into data-frames
}
\details{
Individual EprimeFrames are converted to a data-frame using
  \code{as.data.frame}. (Strings are not converted to factors.)

  Each of the individual data-frames are then \code{rbind}ed together, with
  missing columns being filled with NA.
}
\seealso{
\link{rbind.fill}
}
