% Generated by roxygen2 (4.0.1): do not edit by hand
\name{listify}
\alias{listify}
\title{Convert a vector of colon-separated text lines into a list of named elements}
\usage{
listify(colon_sep_xs)
}
\arguments{
\item{colon_sep_xs}{a character vector with lines of the form
\code{"key: value"}}
}
\value{
a named list of the values in the colon-separated lines.
  \code{"key: value"} yields \code{list(key = "value")}
}
\description{
Convert a vector of colon-separated text lines into a list of named elements
}
\details{
Some minor cleaning of the input is performed:
  \itemize{
    \item Lines without a colon-space separator \code{": "} are filtered out.
    \item Lines with more 400+ characters are filtered out. (This fix is
          directed at \code{Clock.Information:} lines.)
    \item Once the strings are split at the separator, white-space on the
          left and right sides of each half-string is omitted.}
}
\examples{
lines <- c("\\t*** LogFrame Start ***",
           "\\tProcedure: PracticeProc",
           "\\tStimulus1: toothbrush",
           "\\t*** LogFrame End ***")

listify(lines)
# List of 2
# $ Procedure: chr "PracticeProc"
# $ Stimulus1: chr "toothbrush"
}

