% Generated by roxygen2 (4.0.1): do not edit by hand
\name{to_data_frame}
\alias{to_data_frame}
\title{Convert Eprime Frames into data-frames}
\usage{
to_data_frame(...)
}
\arguments{
\item{eprime_frame}{an EprimeFrame object}

\item{frame_list}{a FrameList object (a list of EprimeFrames)}
}
\value{
all of the EprimeFrames combined into a single data frame.
}
\description{
Convert Eprime Frames into data-frames
}
\details{
Individual Eprime Frames are converted to a data-frame using
  \code{as.data.frame}. Strings are not converted to factors.

  Each of the individual data-frames are then \code{rbind}ed together, with
  missing columns being filled with NA.
}
\seealso{
\link{rbind.fill}
}

